% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_voltage_analysis.R
\name{calculate_conduction_mechanisms}
\alias{calculate_conduction_mechanisms}
\title{Calculate Conduction Mechanism Parameters}
\usage{
calculate_conduction_mechanisms(df, voltage_range = c(-1, 1))
}
\arguments{
\item{df}{A data frame containing ferroelectric simulation data}

\item{voltage_range}{Numeric vector of length 2 specifying voltage range for analysis}
}
\value{
A list containing fitted parameters for different conduction mechanisms
}
\description{
This function analyzes the current-voltage relationship to identify different
conduction mechanisms such as ohmic, Schottky emission, and Poole-Frenkel.
}
\examples{
\dontrun{
# Analyze conduction mechanisms in low voltage range
conduction_analysis <- calculate_conduction_mechanisms(data, c(-0.5, 0.5))
}

}
