% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{calculate_parametric_variations}
\alias{calculate_parametric_variations}
\title{Calculate Parametric Variations for Analysis}
\usage{
calculate_parametric_variations(df)
}
\arguments{
\item{df}{A data frame containing processed ferroelectric data}
}
\value{
The input data frame with additional parametric columns:
  \describe{
    \item{freq_1kHz, freq_100kHz}{Frequency analysis components}
    \item{thickness_factor}{Normalized thickness variation factor}
    \item{pol_thickness}{Polarization modified by thickness effects}
    \item{doping_factor}{Doping concentration factor}
    \item{i_leak_doped}{Leakage current modified by doping effects}
  }
}
\description{
This function adds parametric variations to the dataset for studying frequency
response, thickness effects, and doping concentration impacts.
}
\details{
This function simulates the effects of various device parameters:
- Frequency response at 1 kHz and 100 kHz
- Thickness variations and their impact on polarization
- Doping concentration effects on leakage current
}
\examples{
\dontrun{
# Add parametric variations
data_full <- data_processed \%>\%
  calculate_parametric_variations()
}

}
