% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysteresis_analysis.R
\name{compare_hysteresis_loops}
\alias{compare_hysteresis_loops}
\title{Compare Multiple Hysteresis Loops}
\usage{
compare_hysteresis_loops(data_list, labels)
}
\arguments{
\item{data_list}{List of data frames, each containing ferroelectric data}

\item{labels}{Character vector of labels for each dataset}
}
\value{
A ggplot2 object showing multiple hysteresis loops
}
\description{
This function creates a comparison plot of multiple hysteresis loops,
useful for studying parameter variations or different materials.
}
\examples{
\dontrun{
# Compare different conditions
data1 <- process_ferroelectric_data("condition1.csv")
data2 <- process_ferroelectric_data("condition2.csv")
compare_hysteresis_loops(list(data1, data2), c("Condition 1", "Condition 2"))
}

}
