% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric_analysis.R
\name{create_optimization_recommendations}
\alias{create_optimization_recommendations}
\title{Create Parameter Optimization Recommendations}
\usage{
create_optimization_recommendations(sensitivity_results, hysteresis_params)
}
\arguments{
\item{sensitivity_results}{List from analyze_parametric_sensitivity()}

\item{hysteresis_params}{List from extract_hysteresis_parameters()}
}
\value{
A data frame with optimization recommendations
}
\description{
This function analyzes the parametric sensitivity results and provides
recommendations for optimal device design parameters.
}
\examples{
\dontrun{
# Generate optimization recommendations
sensitivity_results <- analyze_parametric_sensitivity(data)
hysteresis_params <- extract_hysteresis_parameters(data)
recommendations <- create_optimization_recommendations(sensitivity_results, hysteresis_params)
print(recommendations)
}

}
